PPM-505: Fix date issues and add questionaire authored date to flattened participant object
For dates with no time, the UTC to ET conversion was pushing the date back a day. Check to see if it is midnight in UTC time and return that date instead of converting to ET.

Also standardizes the string datetime format in the participant object.

Lastly, this adds the authored date to the questionnaire object in the flattened participant object.